+++ pcs_test/tier0/lib/cib/resource/test_stonith.py
@@ -139,6 +139,11 @@ class GetMisconfiguredResources(TestCase):
                         <nvpair name="method" value="cycle" />
                     </instance_attributes>
                 </primitive>
+                <primitive id="S5" class="stonith" type="fence_sbd">
+                    <instance_attributes>
+                        <nvpair name="method" value="cycle" />
+                    </instance_attributes>
+                </primitive>
             </resources>
         """
         )
@@ -146,10 +151,11 @@ class GetMisconfiguredResources(TestCase):
         stonith2 = resources.find("primitive[@id='S2']")
         stonith3 = resources.find("primitive[@id='S3']")
         stonith4 = resources.find("primitive[@id='S4']")
+        stonith5 = resources.find("primitive[@id='S5']")
         self.assertEqual(
             stonith.get_misconfigured_resources(resources),
             (
-                [stonith1, stonith2, stonith3, stonith4],
+                [stonith1, stonith2, stonith3, stonith4, stonith5],
                 [stonith2, stonith4],
                 [stonith3, stonith4],
             ),