+++ CHANGELOG.md
@@ -12,8 +12,11 @@
 ### Fixed
 - Do not end with error when using the instances quantifier in `pcs status
   query resource is-state` command ([RHEL-55441])
+- Do not display a warning in `pcs status` when a fence\_sbd stonith device has
+  its `method` option set to `cycle` ([RHEL-46286])
 
 [RHEL-46284]: https://issues.redhat.com/browse/RHEL-46284
+[RHEL-46286]: https://issues.redhat.com/browse/RHEL-46286
 [RHEL-55441]: https://issues.redhat.com/browse/RHEL-55441